ref(store): updates and deletes
