+++ src/client/src/app/store/index.ts
@@ -6,17 +6,20 @@ import {
   MetaReducer
 } from '@ngrx/store';
 import { environment } from '../../environments/environment';
-import * as fromUser from './reducers/user/user.reducer';
+import * as fromUser from './user/user.reducer';
+import * as fromGame from './game/game.reducer';
 
 
 export interface AppState {
 
   [fromUser.userFeatureKey]: fromUser.UserState;
+  [fromGame.gameFeatureKey]: fromGame.GameState;
 }
 
 export const reducers: ActionReducerMap<AppState> = {
 
   [fromUser.userFeatureKey]: fromUser.reducer,
+  [fromGame.gameFeatureKey]: fromGame.reducer,
 };
