--- src/client/src/app/store/selectors/user/user.selectors.ts
@@ -1,18 +1,0 @@
-import { createFeatureSelector, createSelector } from '@ngrx/store';
-import { AppState } from '../..';
-import * as userFeature from 'src/app/store/reducers/user/user.reducer'
-
-export const userFeatureSelector = createFeatureSelector<AppState, userFeature.UserState>
-(userFeature.userFeatureKey);
-
-//export const userFeatureSelector = (state: AppState) => state.user;
- 
-export const loggedInSelector = createSelector(
-    userFeatureSelector,
-    (state) => state.loggedIn
-);
-
-export const testSelector = createSelector(
-    userFeatureSelector,
-    (state) => state
-);
